backend/requirements.txt
fastapi
uvicorn[standard]
python-docx
pandas
openpyxl
celery
redis
sqlalchemy
alembic
psycopg2-binary
mammoth
